repos/tettettt-main_P0/smart_cane/pitch_yaw_roll/requirements.txt
paho-mqtt
Flask 